Move deploy command into its own folder
